:recycle: 설정 부분 컴포넌트 재사용성을 위해 갈아치움
