+++ client/src/components/room/RoomMenu.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Wrapper, TopBar, CloseButton } from '@components/room/RoomMenu.style';
+import { Wrapper, TopBar, CloseButton, SelectBox } from '@components/room/RoomMenu.style';
 import Chat from '@components/room/chat/';
-import RoomSetting from '@components/setting/RoomSetting';
+import DeviceSelections from '@src/components/setting/DeviceSelections';
 import Host from '@components/room/host/';
 import Users from '@components/user/Users';
 import { useSelector, useDispatch } from 'react-redux';
@@ -16,7 +16,11 @@ const RouteMenu = ({ startVoteRef }) => {
   const { menuType } = useSelector((state: RootState) => state.room);
   switch (menuType) {
     case '설정':
-      return <RoomSetting />;
+      return (
+        <SelectBox>
+          <DeviceSelections />
+        </SelectBox>
+      );
     case '채팅':
       return <Chat />;
     case '참가자':
